+++ CHANGELOG.md
@@ -9,12 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- The `Provenance`, `Publisher`, `AttestationBundle` type have been added
+- The `Provenance`, `Publisher`, `AttestationBundle` types have been added
   ([#36](https://github.com/trailofbits/pypi-attestations/pull/36)).
 
 - The `construct_simple_provenance_object` function has been added, allowing
   to construct a Provenance object from a single publisher kind and a list of
-  attestation from this publisher ([#36](https://github.com/trailofbits/pypi-attestations/pull/36)).
+  attestations from this publisher ([#36](https://github.com/trailofbits/pypi-attestations/pull/36)).
 
 ## [0.0.9]