+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Changed
+
+- The minimum Python version required has been bumped to `3.11`
+  ([#37](https://github.com/trailofbits/pypi-attestations/pull/37))
 
 - The `Provenance`, `Publisher`, `AttestationBundle` types have been added
   ([#36](https://github.com/trailofbits/pypi-attestations/pull/36)).