+++ CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.10]
+
 ### Changed
 
 - The minimum Python version required has been bumped to `3.11`
   ([#37](https://github.com/trailofbits/pypi-attestations/pull/37))
+
+### Added
 
 - The `Provenance`, `Publisher`, `GitHubPublisher`, `GitLabPublisher`, and
   `AttestationBundle` types have been added